requirements.txt
jupyter
kafka-python
psycopg2
pytest
redis
